Change mktables output for some tables to use hex
This makes all the tables in the lib/unicore/To directory that map from
code point to code point be formatted so that the mapped-to code point
is expressed as hexadecimal.

This allows for uniform treatment of these tables in utf8.c, and removes
the final use of strtol() in the (non-CPAN) core.  strtol() should be
avoided because it is subject to locale rules, and some older libc
implementations have been buggy.  It was used because Perl doesn't have
an efficient way of parsing a decimal number and advancing the parse
pointer to beyond it; we do have such a method for hex numbers.

The input to mktables published by Unicode is also in hex, so this now
conforms to that convention.

This also will facilitate the new work currently being done to read in
the tables that find the closing bracket given an opening one.