Remove all IGNORABLE files under cpan/
This was discussed on p5p:
http://www.nntp.perl.org/group/perl.perl5.porters/2013/10/msg208635.html

One LICENSE file (Locale-Codes) has been removed (contrary to one suggestion
in that thread) since it was the only one, and only says "This module is
free software; you can redistribute it and/or modify it under the same
terms as Perl itself." anyway.

Two files that I had previously suggested retaining for the sake of tests
have also been removed. In one case, a test script has been modified (in
Digest-MD5; I will send a ticket upstream); in the other case (in
Test-Harness) I have removed the test which depended on the MANIFEST files,
as suggested in the above thread.