Don't make bool an int on VMS.
The special case has been there since 61bb59065bf1b12edab3, most
likely because the VMS C++ compiler, like a lot of other C++
compilers in the 1990s implemented a bool as an int, and making
the type in C compatible seemed like a good idea.  But no C++
compiler that's likely to build Perl on VMS has a bool type that
occupies more than one byte now, so remove the special case.  We're
unlikely to even see this code since we've had stdbool.h since
DEC C 6.4, released in 2001.