perlfunc: Clarify verbatim table
When I read what was here before this patch, I thought there was a typo,
as there was no entry in column 1 for several rows.  Instead, what was
meant was that the item in column 1 row 1 was supposed to be dittoed
into the following rows.  I used the solution employed in Camel 4,
which is to repeat that item in each row it applies to.